Move configs to different packages
